QGraphicsItem custom tooltip support
This change allows us to show custom tooltips for certain items that
inherit from QGraphicsItem. When the ItemHasCustomTooltip flag is set
sceneEvent will be invoked on the item whenever graphics view decides
that it needs to show a tooltip. The item then may show its own
tooltip. If event->isAccepted() is true (which is the default) no
further processing of the tooltip event will be performed and the
default tooltip will not be shown even if the item has its tooltip()
method return non-empty string().