Disable updating of window when handling specific messages
This is necessary in some cases when composition mode is disabled (or not
available) and certain actions are performed, such as changing the window
icon. Some parts of the default wndproc will paint portions of the titlebar
when it should not.

This locking mechanism works around this by tricking the window into thinking
it isn't visible during the duration of this action. Currently the only handled
action is WM_SETICON.

Change-Id: Id33b21be735a8b283aa668618563bdf5c2fa2153