Allow more control over the order in which renderers are loaded
Added QT_OPENGL_RENDERERS environment variable, which is a comma
separated list of renderers that should be created. If this variable
is set, and a specific renderer is not requested, then attempt
to create each renderer in the order specified.

Valid values are: d3d11, d3d9, warp, angle, desktop, software

Change-Id: I640d6e91b02307dda53a6cc4dbcd369004c894fd